requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
numpy>=1.19.5
Pillow>=8.3.1
tqdm>=4.62.0
requests>=2.26.0
scikit-learn>=0.24.2
